Use xarray for surfacelut and pushes the read to static method in surfacelut
